requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.20
alembic>=1.13.0
psycopg2-binary>=2.9.7
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.6.0
chromadb>=0.4.15
PyPDF2>=3.0.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
requests>=2.31.0
reportlab>=4.0.0